+++ src/components/Banner/index.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { Container } from 'react-bootstrap'
 import './index.less'
