+++ src/components/ListItem/index.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { Button } from 'react-bootstrap'
 
 import { BASE_URL } from '../../constants/settiings'
