+++ src/components/ListView/index.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { useNavigate } from 'react-router-dom'
 import ListItem from '../ListItem'
 import './index.less'
