+++ src/pages/TopicDetail/CommentForm/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import { useState } from 'react'
 import { Button, Card, Form, Spinner } from 'react-bootstrap'
 
 import { useAcountStore } from '../../../stores/auth'
